feat WishService.java, WishController.java: 위시리스트 기능 구현

+++ src/main/java/gift/jwt/JwtService.java
@@ -34,7 +34,7 @@ public class JwtService {
     }
 
     public Long getUserId(String token){
-        return extractClaims(token).get("userId", Long.class);
+        return extractClaims(token).get("userId", Double.class).longValue();
     }
 
     public String getUserName(String token){
